move tests from advz.rs to new file advz/test.rs
